PLAT-1669 : abac ship UI , random board (#64)
* ЗДФЕ-1669 : random board


* ЗДФЕ-1669 : reveal cell fix
