+++ examples/abacship-app/src/containers/GameDesk/utils.ts
@@ -1,7 +1,8 @@
 import { ASSETS_LIST } from "../../assets";
-import { CELL_TYPE } from "../../models/cellType";
-import { getTDFData, TDFDATA } from "../../models/tdfs";
-import { postBoard } from "../../services/axios";
+import { BoardKeyType, SecretBoard, ShipType, StatusBoard } from "../../interfaces/board";
+import { CELL_TYPE, COL_INDICATORS, ROW_INDICATORS } from "../../models/cellType";
+import { defaulBoardSize, getBoard, setBoard } from "../../utils/board";
+import { generateRandomBoard } from "../../utils/randomBoard";
 
 const { SOUNDS } = ASSETS_LIST;
 
@@ -15,44 +16,57 @@ export function getAudioForCell(cell: number) {
   return null;
 }
 
-export function revealCell(rowIdx: number, colIdx: number) {
-  // actual hit
-  const data = getTDFData(rowIdx, colIdx);
-
-  if (data === "ocean") return CELL_TYPE.OCEAN;
-  if (data === "aircraft carrier" || data === "battleship" ||
-  data === "cruiser" || data === "destroyer" ||
-  data === "submarine") return CELL_TYPE.PLAYER_ONE;
-  if (data === "aircraft carrier" || data === "battleship" ||
-  data === "cruiser" || data === "destroyer" ||
-  data === "submarine") return CELL_TYPE.PLAYER_TWO; // yeah , i know , prob i'll delete this
+export function revealCell(value: string) {
+  if (value === "ocean") return CELL_TYPE.OCEAN;
+  if (Object.values(ShipType).includes(value as unknown as ShipType)) return CELL_TYPE.PLAYER_ONE;
+  // if (Object.values(ShipType).includes(data as unknown as ShipType)) return CELL_TYPE.PLAYER_TWO; // todo player two
 }
 
-export const hitGridItem = (myGrid: number[][], rowIdx: number, colIdx: number) => {
-  const newGridRows: number[][] = [];
-  myGrid.forEach((oldRow: any[], oldRowIdx: number) => {
-    const newGridRow: number[] = [];
-    oldRow.forEach((oldCell, oldColIdx) => {
+export const hitGridItem = (statusBoard: number[][], rowIdx: number, colIdx: number, secretValue: string) => {
+  return statusBoard.map((oldRow: number[], oldRowIdx: number) => {
+    return oldRow.map((oldCell, oldColIdx) => {
       if (rowIdx === oldRowIdx && colIdx === oldColIdx) {
         // This is the cell we are revealing.
-        const cell = revealCell(rowIdx, colIdx);
+        // const _secret_board = getBoard<>();
+        const cell = revealCell(secretValue);
         const audioForCell = getAudioForCell(cell);
         if (audioForCell !== null) {
           audioForCell.play();
         }
-        newGridRow.push(cell);
+        return cell;
       } else {
-        newGridRow.push(oldCell);
+        return oldCell;
       }
     });
-    newGridRows.push(newGridRow);
   });
-
-  return newGridRows;
 };
 
-export const sendBoard = async (player: string) => {
-  const token = sessionStorage.getItem("token") || "";
-  const refreshToken = sessionStorage.getItem("refreshToken") || "";
-  postBoard(token, refreshToken, player, TDFDATA);
-};
+export const getMyGrid = async (): Promise<number[][]> => {
+  const boardKey = "my_secret_board";
+  const localBoard = getBoard<SecretBoard>(boardKey);
+  let board: SecretBoard;
+  if (localBoard.length === 0) {
+    const randomBoard = generateRandomBoard(defaulBoardSize);
+    setBoard(boardKey, randomBoard);
+    board = randomBoard;
+  } else {
+    board = localBoard;
+  }
+
+  return board.map((row, rowIdx) => row.map((cell, colIdx) => revealCell(cell)));
+}
+
+export const getOpponentGrid = async (): Promise<number[][]> => {
+  const boardKey = "enemy_board";
+  const localBoard = getBoard<StatusBoard>(boardKey);
+  const emptyBoard = ROW_INDICATORS.map(() => COL_INDICATORS.map(() => CELL_TYPE.UNKNOWN));
+  let board;
+  if (localBoard.length === 0) {
+    setBoard(boardKey, emptyBoard);
+    board = emptyBoard;
+  } else {
+    board = localBoard;
+  }
+
+  return board;
+}
