+++ examples/abacship-app/src/models/cellType.ts
@@ -5,5 +5,6 @@ export const CELL_TYPE = {
   UNKNOWN: 3,
 };
 
+
 export const ROW_INDICATORS = ["A", "B", "C", "D", "E", "F", "G", "H", "I", "J"];
 export const COL_INDICATORS = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10"];
